move chunkrequest to separate file
